move test to unit tests and resolve overriding method
